fixed utils
