+++ packages/domql/index.js
@@ -2,7 +2,7 @@
 
 import { window } from '@domql/globals'
 import { TREE } from '@domql/tree'
-import { create, parse, set, define } from './element'
+import { create, parse, set, define } from './src'
 import { isDevelopment, isTest } from '@domql/env'
 
 if (isTest() || isDevelopment()) window.tree = TREE
