+++ packages/domql/package.json
@@ -28,10 +28,12 @@
     "@domql/env": "latest",
     "@domql/node": "latest",
     "@domql/render": "latest",
+    "@domql/props": "latest",
     "@domql/tree": "latest",
     "@domql/globals": "latest",
     "@domql/registry": "latest",
     "@domql/methods": "latest",
+    "@domql/state": "latest",
     "@domql/event": "latest",
     "@domql/report": "latest",
     "@domql/emotion": "latest",
