+++ packages/domql/src/node.js
@@ -11,6 +11,7 @@ import {
   throughInitialDefine,
   throughInitialExec
 } from './iterate'
+
 import { registry } from './mixins'
 import { applyParam } from './applyParam'
 // import { defineSetter } from './methods'
