+++ packages/props/index.js
@@ -1,63 +1,109 @@
 'use strict'
 
-import { deepMerge, exec, isArray } from '@domql/utils'
+import {
+  exec,
+  is,
+  isArray,
+  isObject,
+  isString,
+  deepClone,
+  deepMerge
+} from '@domql/utils'
 
-const initPropsStack = (element, parent) => {
-  const propsStack = []
+export const IGNORE_PROPS_PARAMS = ['update', '__element']
 
-  if (element.props === 'inherit') {
-    if (parent && parent.props) propsStack.push(parent.props)
-  } else if (element.props === 'match') {
-    if (parent && parent.props) propsStack.push(parent.props[element.key])
-  } else if (element.props) propsStack.push(element.props)
+const objectizeStringProperty = propValue => {
+  if (is(propValue)('string', 'number')) return { inheritedString: propValue }
+  return propValue
+}
 
-  if (isArray(element.ref.__extend)) {
-    element.ref.__extend.map(extendUnit => {
-      if (extendUnit.props) propsStack.push(extendUnit.props)
-      return extendUnit.props
-    })
+const inheritParentProps = (element, parent) => {
+  let propsStack = []
+  const parentProps = exec(parent, parent.state).props
+
+  const matchParent = parent.props && parentProps[element.key]
+  const matchParentIsString = isString(matchParent)
+  const matchParentChildProps = parentProps && parentProps.childProps
+
+  if (matchParent) {
+    if (matchParentIsString) {
+      const inheritedStringExists = propsStack.filter(v => v.inheritedString)[0]
+      if (inheritedStringExists) inheritedStringExists.inheritedString = matchParent
+      else {
+        propsStack = [].concat(objectizeStringProperty(matchParent), propsStack)
+      }
+    } else {
+      propsStack.push(objectizeStringProperty(matchParent))
+    }
   }
+  if (matchParentChildProps) propsStack.push(matchParentChildProps)
 
   return propsStack
 }
 
-export const inheritProps = (element, parent) => {
-  return (parent && parent.ref.props)
-}
+const createPropsStack = (element, parent) => {
+  const { props, __ref } = element
+  const propsStack = __ref.__props = inheritParentProps(element, parent)
 
-export const syncProps = (propsStack, element) => {
-  const { ref } = element
-  const mergedProps = {}
-  propsStack.forEach(prop => {
-    // to realtime sync props for lazy exec
-    ref.props = deepMerge(mergedProps, exec(prop, element))
-  })
-  return mergedProps
-}
+  if (isObject(props)) propsStack.push(props)
+  else if (props === 'inherit' && parent.props) propsStack.push(parent.props)
+  else if (props) propsStack.push(props)
 
-export const createProps = function (element, parent, cache) {
-  const { ref } = element
-  if (!ref.props) ref.props = {}
-  const propsStack = cache || initPropsStack(element, parent)
-
-  if (propsStack.length) {
-    ref.__props = propsStack
-    return syncProps(propsStack, element)
+  if (isArray(__ref.__extend)) {
+    __ref.__extend.map(extend => {
+      if (extend.props) propsStack.push(extend.props)
+      return extend.props
+    })
   }
 
-  return inheritProps(element, parent)
+  __ref.__props = propsStack
+
+  return propsStack
+}
+
+export const syncProps = (props, element) => {
+  element.props = {}
+  const mergedProps = { update, __element: element }
+  props.forEach(v => {
+    if (IGNORE_PROPS_PARAMS.includes(v)) return
+    const execProps = exec(v, element)
+    if (isObject(execProps) && execProps.__element) return
+    element.props = deepMerge(mergedProps, deepClone(execProps, IGNORE_PROPS_PARAMS), IGNORE_PROPS_PARAMS)
+  })
+  element.props = mergedProps
+  return element.props
+}
+
+export const createProps = function (element, parent, cached) {
+  const propsStack = cached || createPropsStack(element, parent)
+  const { __ref } = element
+
+  if (propsStack.length) {
+    __ref.__props = propsStack
+    syncProps(propsStack, element)
+    element.props.update = update
+  }
+
+  return element
 }
 
 export const updateProps = (newProps, element, parent) => {
-  const { ref } = element
-  let propsStack = ref.__props
+  const { __ref } = element
+  let propsStack = __ref.__props
 
-  if (newProps) propsStack = [].concat(newProps, propsStack)
+  const parentProps = inheritParentProps(element, parent)
+  if (parentProps) propsStack = __ref.__props = [].concat(parentProps, propsStack)
+  if (newProps) propsStack = __ref.__props = [].concat(newProps, propsStack)
 
-  if (propsStack.length) {
-    ref.__props = propsStack
-    return syncProps(propsStack, element)
-  }
+  if (propsStack) syncProps(propsStack, element)
 
-  return inheritProps(element, parent)
+  // console.log(cachedProps)
+  return element
 }
+
+function update (props, options) {
+  const element = this.__element
+  element.update({ props }, options)
+}
+
+export default createProps
